pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crispex"
version = "0.1.0"
description = "AI-powered CRISPR sgRNA design with ensemble ML and off-target detection"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Crispex Contributors"}
]
keywords = ["crispr", "sgrna", "guide-design", "gene-editing", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "biopython>=1.79",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "requests>=2.26.0",
    "click>=8.0.0",
    "pyfaidx>=0.7.0",
    "tqdm>=4.62.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.scripts]
crispex = "crispex.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/crispex"
Repository = "https://github.com/yourusername/crispex"
Documentation = "https://github.com/yourusername/crispex#readme"
"Bug Tracker" = "https://github.com/yourusername/crispex/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["crispex*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
crispex = ["data/models/*.pkl", "data/configs/*.json"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=crispex --cov-report=term-missing"
